Feat. change changedDoc state to atom via jotai and move fetch function for debug
